+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 Livedata SDK is a client library that enables easier integration with the Livescout XML feed. SDK exposes XML feed service interface in a more user-friendly way and isolates the client from having to do XML feed parsing, proper connection handling, error recovery, event queuing and dispatching. 
 It also makes a client solution more stable and robust when it comes to feed handling, especially with the release of new and updated XML feed version.* ScoutFeedType extended with new type - PARTIAL
 
+**2.0.6 (2023-11-08)**
+* MatchHeaderEntity extended with new property: region
+* MatchUpdateEntity subteam -> subteams
+* ScoutEventEntity extended with new properties:
+  - homePlayerStats
+  - awayPlayerStats
+
 **2.0.5 (2023-09-20)**
 * FormatType extended with new type - USES_OVERS
 * MatchHeaderEntity getters names changed: setIsCancelled -> isCancelled, setIsTeamMatch -> isTeamMatch